requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
apscheduler>=3.10.0
loguru>=0.7.0
pandas>=2.2.0
pyyaml>=6.0
ruamel.yaml>=0.18.0
pymongo>=4.0.0
jinja2>=3.1.0
